+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Date: Month DD, 2022
 URL: <https://github.com/KCarlile/flashchord/releases/tag/v1.8>  
 Commit: <>
 
-- ...next change.
+- Fixed bug in debugging-enabled check (no issue).
+- 
 
 ## Release 1.7