+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Commit: <>
 - Issue [#60](https://github.com/KCarlile/flashchord/issues/60): Add changelog to About page.
 - About page updates: adding new tech and adding donors.
 - Issue [#80](https://github.com/KCarlile/flashchord/issues/80): Updated to Twitter Bootstrap 5.2-beta.
+- Issue [#73](https://github.com/KCarlile/flashchord/issues/73): Add Jest test suite for unit/regression testing.
 
 ## Release 1.7